# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Nifty100 Analytics API","description":"REST API for Nifty100 Analytics Platform","version":"1.0.0"},"paths":{"/api/v1/clusters":{"get":{"tags":["Clusters"],"summary":"Get clustered companies","description":"Returns clustered companies with optional filtering.","operationId":"get_clusters_api_v1_clusters_get","parameters":[{"name":"cluster_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"description":"Filter by cluster ID","title":"Cluster Id"},"description":"Filter by cluster ID"},{"name":"sector","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by broad sector","title":"Sector"},"description":"Filter by broad sector"},{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Search by company ID or company name","title":"Search"},"description":"Search by company ID or company name"},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"Page number","default":1,"title":"Page"},"description":"Page number"},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"Number of records per page","default":10,"title":"Page Size"},"description":"Number of records per page"},{"name":"sort_by","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Column to sort by","title":"Sort By"},"description":"Column to sort by"},{"name":"order","in":"query","required":false,"schema":{"type":"string","pattern":"^(asc|desc)$","description":"Sort order: asc or desc","default":"asc","title":"Order"},"description":"Sort order: asc or desc"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClusterListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/clusters/report":{"get":{"tags":["Clusters"],"summary":"Cluster summary report","description":"Returns aggregated statistics for each cluster.","operationId":"get_cluster_report_api_v1_clusters_report_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ClusterReport"},"type":"array","title":"Response Get Cluster Report Api V1 Clusters Report Get"}}}}}}},"/api/v1/clusters/profiles":{"get":{"tags":["Clusters"],"summary":"Cluster feature profiles","description":"Returns the average value of each feature for every cluster.","operationId":"get_cluster_profiles_api_v1_clusters_profiles_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ClusterProfile"},"type":"array","title":"Response Get Cluster Profiles Api V1 Clusters Profiles Get"}}}}}}},"/api/v1/clusters/sectors":{"get":{"tags":["Clusters"],"summary":"Cluster sector distribution","description":"Returns the sector-wise company distribution for every cluster.","operationId":"get_cluster_sector_distribution_api_v1_clusters_sectors_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ClusterSectorDistribution"},"type":"array","title":"Response Get Cluster Sector Distribution Api V1 Clusters Sectors Get"}}}}}}},"/api/v1/companies":{"get":{"tags":["Companies"],"summary":"Get Companies","operationId":"get_companies_api_v1_companies_get","parameters":[{"name":"sector","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by broad sector","title":"Sector"},"description":"Filter by broad sector"},{"name":"market_cap_category","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by market cap category","title":"Market Cap Category"},"description":"Filter by market cap category"},{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Search by company name or ticker","title":"Search"},"description":"Search by company name or ticker"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}":{"get":{"tags":["Companies"],"summary":"Get Company Profile","operationId":"get_company_profile_api_v1_companies__ticker__get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/pl":{"get":{"tags":["Companies"],"summary":"Get Profit Loss","operationId":"get_profit_loss_api_v1_companies__ticker__pl_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}},{"name":"from_year","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Start period in YYYY-MM format","title":"From Year"},"description":"Start period in YYYY-MM format"},{"name":"to_year","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"End period in YYYY-MM format","title":"To Year"},"description":"End period in YYYY-MM format"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/bs":{"get":{"tags":["Companies"],"summary":"Get Balance Sheet","operationId":"get_balance_sheet_api_v1_companies__ticker__bs_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}},{"name":"from_year","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Start period in YYYY-MM format","title":"From Year"},"description":"Start period in YYYY-MM format"},{"name":"to_year","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"End period in YYYY-MM format","title":"To Year"},"description":"End period in YYYY-MM format"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/cashflow":{"get":{"tags":["Companies"],"summary":"Get Cashflow","operationId":"get_cashflow_api_v1_companies__ticker__cashflow_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}},{"name":"from_year","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Start period in YYYY-MM format","title":"From Year"},"description":"Start period in YYYY-MM format"},{"name":"to_year","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"End period in YYYY-MM format","title":"To Year"},"description":"End period in YYYY-MM format"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/ratios":{"get":{"tags":["Companies"],"summary":"Get Ratios","operationId":"get_ratios_api_v1_companies__ticker__ratios_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}},{"name":"year","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Year"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/tearsheet":{"get":{"tags":["Companies"],"summary":"Get Tearsheet","operationId":"get_tearsheet_api_v1_companies__ticker__tearsheet_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/screener":{"get":{"tags":["Screener"],"summary":"Screener","operationId":"screener_api_v1_screener_get","parameters":[{"name":"min_roe","in":"query","required":false,"schema":{"anyOf":[{"type":"number","minimum":0},{"type":"null"}],"title":"Min Roe"}},{"name":"max_de","in":"query","required":false,"schema":{"anyOf":[{"type":"number","minimum":0},{"type":"null"}],"title":"Max De"}},{"name":"min_fcf","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Min Fcf"}},{"name":"sector","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sector"}},{"name":"min_rev_cagr_5yr","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Min Rev Cagr 5Yr"}},{"name":"min_pat_cagr_5yr","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Min Pat Cagr 5Yr"}},{"name":"max_pe","in":"query","required":false,"schema":{"anyOf":[{"type":"number","exclusiveMinimum":0},{"type":"null"}],"title":"Max Pe"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/sectors/{sector}/companies":{"get":{"tags":["Sectors"],"summary":"Get Sector Companies","operationId":"get_sector_companies_api_v1_sectors__sector__companies_get","parameters":[{"name":"sector","in":"path","required":true,"schema":{"type":"string","title":"Sector"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{ticker}/peers/compare":{"get":{"tags":["Peers"],"summary":"Compare With Peers","operationId":"compare_with_peers_api_v1_companies__ticker__peers_compare_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/market-cap/{ticker}":{"get":{"tags":["Valuation"],"summary":"Get Market Cap History","operationId":"get_market_cap_history_api_v1_market_cap__ticker__get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/portfolio/stats":{"get":{"tags":["Portfolio"],"summary":"Portfolio Stats","operationId":"portfolio_stats_api_v1_portfolio_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/companies/{ticker}/documents":{"get":{"tags":["Documents"],"summary":"Get Documents","operationId":"get_documents_api_v1_companies__ticker__documents_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/health":{"get":{"tags":["Health"],"summary":"Health Check","operationId":"health_check_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/":{"get":{"tags":["Default"],"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Cluster":{"properties":{"company_id":{"type":"string","title":"Company Id"},"company_name":{"type":"string","title":"Company Name"},"broad_sector":{"type":"string","title":"Broad Sector"},"cluster_id":{"type":"integer","title":"Cluster Id"},"cluster_name":{"type":"string","title":"Cluster Name"},"distance_from_centroid":{"type":"number","title":"Distance From Centroid"}},"type":"object","required":["company_id","company_name","broad_sector","cluster_id","cluster_name","distance_from_centroid"],"title":"Cluster"},"ClusterListResponse":{"properties":{"page":{"type":"integer","title":"Page"},"page_size":{"type":"integer","title":"Page Size"},"total_records":{"type":"integer","title":"Total Records"},"total_pages":{"type":"integer","title":"Total Pages"},"data":{"items":{"$ref":"#/components/schemas/Cluster"},"type":"array","title":"Data"}},"type":"object","required":["page","page_size","total_records","total_pages","data"],"title":"ClusterListResponse"},"ClusterProfile":{"properties":{"cluster_id":{"type":"integer","title":"Cluster Id"},"feature":{"type":"string","title":"Feature"},"mean_value":{"type":"number","title":"Mean Value"}},"type":"object","required":["cluster_id","feature","mean_value"],"title":"ClusterProfile"},"ClusterReport":{"properties":{"cluster_id":{"type":"integer","title":"Cluster Id"},"company_count":{"type":"integer","title":"Company Count"},"cluster_name":{"type":"string","title":"Cluster Name"},"dominant_sector":{"type":"string","title":"Dominant Sector"},"return_on_equity_pct":{"type":"number","title":"Return On Equity Pct"},"debt_to_equity":{"type":"number","title":"Debt To Equity"},"revenue_cagr_5yr":{"type":"number","title":"Revenue Cagr 5Yr"},"operating_profit_margin_pct":{"type":"number","title":"Operating Profit Margin Pct"},"free_cash_flow_cr":{"type":"number","title":"Free Cash Flow Cr"}},"type":"object","required":["cluster_id","company_count","cluster_name","dominant_sector","return_on_equity_pct","debt_to_equity","revenue_cagr_5yr","operating_profit_margin_pct","free_cash_flow_cr"],"title":"ClusterReport"},"ClusterSectorDistribution":{"properties":{"cluster_id":{"type":"integer","title":"Cluster Id"},"broad_sector":{"type":"string","title":"Broad Sector"},"company_count":{"type":"integer","title":"Company Count"}},"type":"object","required":["cluster_id","broad_sector","company_count"],"title":"ClusterSectorDistribution"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}